move settings tests to its own file
